refactor(): BIG structure change * moved all 'core' things to a new core folder * reorganized / moved other pieces around * updated all require() calls
