+++ src/core/FuzzyFactory.js
