+++ src/core/FuzzyFactory2d.js
