+++ src/core/FuzzyFactory3d.js
