+++ src/core/Properties.js
