+++ src/core/math/Path2.js
