+++ src/core/math/Polygon2.js
