+++ src/core/math/Vector3.js
