+++ src/core/math/Vertex2.js
