+++ src/core/math/Vertex3.js
