+++ src/core/math/polygonUtils.js
