+++ src/core/trees.js
