+++ src/core/utils/cagMeasurements.js
