+++ src/core/utils/canonicalize.js
