+++ src/core/utils/isCAGValid.js
